refactor(editor): :art: Change eidotor directory names
- ArticleEditor -> inputs
- Change directory name to lowercase
