+++ client/src/components/common/Editor/Inputs/__test__/InputWrapper.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import InputWrapper from "./InputWrapper";
+import InputWrapper from "../InputWrapper";
 
 describe("<InpurWrapper />", () => {
   it("renders its name and children", () => {
