+++ client/src/components/common/Editor/Inputs/index.ts
@@ -1,0 +1,4 @@
+export { default as InputWrapper } from "./InputWrapper";
+export { default as SourceInput } from "./SourceInput";
+export { default as ContentInput } from "./ContentInput";
+export { default as TagInput } from "./TagInput";
